--- baseDecorator.rb
@@ -1,14 +1,0 @@
-require './nameable'
-
-class BaseDecorator < Nameable
-  attr_accessor :nameable
-
-    def initialize(nameable)
-        super()
-        @nameable = nameable
-    end
-
-    def correct_name
-        @nameable.correct_name
-    end
-end
