+++ capitalize_decorator.rb
@@ -1,0 +1,7 @@
+require './base_decorator'
+
+class capitalizedPerson < baseDecorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
