+++ person.rb
@@ -5,7 +5,6 @@ class Person < Nameable
   attr_accessor :name, :age
 
   def initialize(age, name: 'unknown', parent_permission: true)
-
     super()
 
     @id = Random.rand(1..1000)
@@ -25,5 +24,4 @@ class Person < Nameable
   def correct_name
     @name
   end
-
 end
