+++ trimmer_decorator.rb
@@ -1,0 +1,8 @@
+require './base_decorator'
+class TrimmerDecorator < baseDecorator
+  def correct_name
+    return @nameable.correct_name[0..9] if @nameable.correct_name.length < 10
+
+    @nameable.correct_name
+  end
+end
